Load same index.js with iOS and Android
